Updated the remodeling backup arguments

+++ tests/tools/remodeling/cli/test_run_remodel_backup.py
@@ -15,7 +15,9 @@ class Test(unittest.TestCase):
     def setUpClass(cls):
         file_list = ['top_level.tsv', 'sub1/sub1_events.tsv', 'sub2/sub2_events.tsv', 'sub2/sub2_next_events.tsv']
         # cls.file_list = file_list
-        cls.extract_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../../../data/remodel_tests')
+        extract_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../../../data/remodel_tests')
+        cls.alt_path = os.path.realpath(os.path.join(extract_path, 'temp'))
+        cls.extract_path = extract_path
         test_root = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../../../data/remodel_tests/test_root')
         cls.test_root = test_root
         cls.test_paths = [os.path.join(test_root, file) for file in file_list]
@@ -34,8 +36,12 @@ class Test(unittest.TestCase):
             zip_ref.extractall(self.extract_path)
 
     def tearDown(self):
-        shutil.rmtree(self.test_root)
-        shutil.rmtree(self.data_root)
+        if os.path.exists(self.test_root):
+            shutil.rmtree(self.test_root)
+        if os.path.exists(self.data_root):
+            shutil.rmtree(self.data_root)
+        if os.path.exists(self.alt_path):
+            shutil.rmtree(self.alt_path)
 
     @classmethod
     def tearDownClass(cls):
@@ -43,11 +49,11 @@ class Test(unittest.TestCase):
 
     def test_main_events(self):
         self.assertFalse(os.path.exists(self.derv_path), 'backup directory does not exist before creation')
-        arg_list = [self.test_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
+        arg_list = [self.test_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-bd', self.derv_path, '-x', 'derivatives',
                     '-f', 'events', '-e', '.tsv']
         main(arg_list)
         self.assertTrue(os.path.exists(self.derv_path), 'backup directory exists before creation')
-        json_path = os.path.realpath(os.path.join(self.derv_path, 'backups', BackupManager.DEFAULT_BACKUP_NAME,
+        json_path = os.path.realpath(os.path.join(self.derv_path, BackupManager.DEFAULT_BACKUP_NAME,
                                                   BackupManager.BACKUP_DICTIONARY))
         with open(json_path, 'r') as fp:
             key_dict = json.load(fp)
@@ -56,19 +62,18 @@ class Test(unittest.TestCase):
         self.assertEqual(len(file_list), 3, "The backup of events.tsv has the right number of files")
 
     def test_main_all(self):
-        arg_list = [self.test_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
-                    '-f', '*', '-e', '*']
+        arg_list = [self.test_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-bd', self.derv_path,
+                    '-x', 'derivatives', '-f', '*', '-e', '*']
 
         self.assertFalse(os.path.exists(self.derv_path), 'backup directory does not exist before creation')
         main(arg_list)
         self.assertTrue(os.path.exists(self.derv_path), 'backup directory exists before creation')
-        json_path = os.path.realpath(os.path.join(self.derv_path, 'backups', BackupManager.DEFAULT_BACKUP_NAME,
+        json_path = os.path.realpath(os.path.join(self.derv_path, BackupManager.DEFAULT_BACKUP_NAME,
                                                   BackupManager.BACKUP_DICTIONARY))
         with open(json_path, 'r') as fp:
             key_dict = json.load(fp)
         self.assertEqual(len(key_dict), 4, "The backup of events.tsv does not include top_level.tsv")
-        back_path = os.path.realpath(os.path.join(self.derv_path, 'backups', 
-                                                  BackupManager.DEFAULT_BACKUP_NAME, 'backup_root'))
+        back_path = os.path.realpath(os.path.join(self.derv_path, BackupManager.DEFAULT_BACKUP_NAME, 'backup_root'))
         file_list1 = get_file_list(back_path)
         self.assertIsInstance(file_list1, list)
         self.assertEqual(len(file_list1), 4)
@@ -78,11 +83,11 @@ class Test(unittest.TestCase):
         self.assertTrue(os.path.exists(der_path))
         shutil.rmtree(der_path)
         self.assertFalse(os.path.exists(der_path))
-        arg_list = [self.data_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
+        arg_list = [self.data_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
                     '-f', 'events', '-e', '.tsv', '-t', 'FacePerception']
         main(arg_list)
         self.assertTrue(os.path.exists(der_path))
-        back_path = os.path.realpath(os.path.join(self.data_root, BackupManager.RELATIVE_BACKUP_LOCATION, 'backups',
+        back_path = os.path.realpath(os.path.join(self.data_root, BackupManager.RELATIVE_BACKUP_LOCATION,
                                                   BackupManager.DEFAULT_BACKUP_NAME, 'backup_root'))
         self.assertTrue(os.path.exists(back_path))
         backed_files = get_file_list(back_path)
@@ -93,37 +98,33 @@ class Test(unittest.TestCase):
         self.assertTrue(os.path.exists(der_path))
         shutil.rmtree(der_path)
         self.assertFalse(os.path.exists(der_path))
-        arg_list = [self.data_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
+        arg_list = [self.data_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
                     '-f', 'events', '-e', '.tsv', '-t', 'Baloney']
         main(arg_list)
         self.assertTrue(os.path.exists(der_path))
-        back_path = os.path.realpath(os.path.join(self.data_root, BackupManager.RELATIVE_BACKUP_LOCATION, 'backups',
+        back_path = os.path.realpath(os.path.join(self.data_root, BackupManager.RELATIVE_BACKUP_LOCATION, 
                                                   BackupManager.DEFAULT_BACKUP_NAME, 'backup_root'))
         self.assertTrue(os.path.exists(back_path))
         backed_files = get_file_list(back_path)
         self.assertEqual(len(backed_files), 0)
 
     def test_alt_loc(self):
-        alt_path = os.path.realpath(os.path.join(self.extract_path, 'temp'))
-        if os.path.exists(alt_path):
-            shutil.rmtree(alt_path)
-        self.assertFalse(os.path.exists(alt_path))
-        arg_list = [self.data_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives', '-w', alt_path,
+        if os.path.exists(self.alt_path):
+            shutil.rmtree(self.alt_path)
+        self.assertFalse(os.path.exists(self.alt_path))
+        arg_list = [self.data_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives', '-bd', self.alt_path,
                     '-f', 'events', '-e', '.tsv', ]
         main(arg_list)
-        self.assertTrue(os.path.exists(alt_path))
-        back_path = os.path.realpath(os.path.join(alt_path, 'backups', 'default_back', 'backup_root'))
-        self.assertTrue(os.path.exists(back_path))
+        self.assertTrue(os.path.exists(self.alt_path))
+        back_path = os.path.realpath(os.path.join(self.alt_path, 'default_back/backup_root'))
         self.assertTrue(os.path.exists(back_path))
         backed_files = get_file_list(back_path)
         self.assertEqual(len(backed_files), 6)
-        if os.path.exists(alt_path):
-            shutil.rmtree(alt_path)
 
     def test_main_backup_exists(self):
         der_path = os.path.realpath(os.path.join(self.data_root, 'derivatives'))
         self.assertTrue(os.path.exists(der_path))
-        arg_list = [self.data_root, '-n', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
+        arg_list = [self.data_root, '-bn', BackupManager.DEFAULT_BACKUP_NAME, '-x', 'derivatives',
                     '-f', 'events', '-e', '.tsv', '-t', 'Baloney']
         with self.assertRaises(HedFileError) as context:
             main(arg_list)
